[CP-403] Replace old update code
Moved Update request to api/update.py
